pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notebook-parser"
version = "0.1.0"
description = "CLI tool for extracting text from images using OCR"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "pillow>=10.0.0",
    "easyocr>=1.7.2",
    "transformers>=4.57.6",
    "torch>=2.10.0",
    "opencv-python>=4.13.0.90",
    "numpy>=2.4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
notebook-parser = "notebook_parser.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing"
